Reorganize modal files
