+++ webui/react/src/pages/WorkspaceDetails/ProjectCard.tsx
@@ -6,9 +6,9 @@ import Avatar from 'components/Avatar';
 import Icon from 'components/Icon';
 import Link from 'components/Link';
 import TimeAgo from 'components/TimeAgo';
-import useModalProjectDelete from 'hooks/useModal/useModalProjectDelete';
-import useModalProjectEdit from 'hooks/useModal/useModalProjectEdit';
-import useModalProjectMove from 'hooks/useModal/useModalProjectMove';
+import useModalProjectDelete from 'hooks/useModal/Project/useModalProjectDelete';
+import useModalProjectEdit from 'hooks/useModal/Project/useModalProjectEdit';
+import useModalProjectMove from 'hooks/useModal/Project/useModalProjectMove';
 import { paths } from 'routes/utils';
 import { archiveProject, unarchiveProject } from 'services/api';
 import { Project } from 'types';
